VAULT-31181: Add `pipeline` tool to Vault
As the Vault pipeline and release processes evolves over time, so too must
the tooling that drives it. Historically we've utilized a combination of CI
features and shell scripts that are wrapped into make targets to drive
our CI. While this approach has worked, it requires careful consideration
of what features to use (bash in CI almost never matches bash in developer
machines, etc.) and often requires a deep understanding of several CLI
tools (jq, etc). `make` itself also has limitations in user experience, e.g.
passing flags.

As we're all in on Github Actions as our pipeline coordinator, continuing to
utilize and build CLI tools to perform our pipeline tasks makes sense.
This PR adds a new CLI tool called `pipeline` which we can use to build
new isolated tasks that we can string together in Github Actions. We intend
to use this utility as the interface for future release automation work,
see VAULT-27514.

For the first task in this new pipeline tool, I've chosen to build a
small command that allows us to list Vault versions between a range.
The goal is to utilize this tool on all active branches to automatically
populate a list of all N-3 versions relative to that Vault edition and
branch. We'll feed that list of released versions into our Enos sampling
algorithm and make all of those versions candidates in our upgrade
testing scenarios, allowing us to test upgrades from all supported
versions over the lifecycle of a given Vault version. As it's automated
and dyanmic we will no longer have to maintain a list of initial
versions on each branch.

We also create a new tools Go module so that our pipeline tooling
dependencies will not be required to build Vault.

Signed-off-by: Ryan Cragun <me@ryan.ec>